feat: parallelized requests, refactored whole code
